Use behavioral HLS recovery coverage
